feat: Refactor Comment handling to use Long IDs, implement CommentResponse DTO, and update related services and controllers

+++ src/main/java/org/amalitech/bloggingplatformspring/graphql/resolvers/GraphQLQueryResolver.java
@@ -102,7 +102,7 @@ public class GraphQLQueryResolver {
     }
 
     @QueryMapping
-    public List<GraphQLComment> getCommentsByPost(@Argument Integer postId) {
+    public List<GraphQLComment> getCommentsByPost(@Argument Long postId) {
         List<CommentResponse> comments = commentService.getAllCommentsByPostId(postId);
         return comments.stream()
                 .map(this::mapToGraphQLComment)
